+++ src/commands/validate.rs
@@ -5,6 +5,7 @@ use anyhow::anyhow;
 use clap::Parser;
 use wdl::engine::Inputs as EngineInputs;
 
+use crate::Config;
 use crate::analysis::Analysis;
 use crate::analysis::Source;
 use crate::commands::CommandError;
@@ -48,18 +49,8 @@ pub struct Args {
     pub report_mode: Option<Mode>,
 }
 
-impl Args {
-    /// Applies the configuration to the arguments.
-    pub fn apply(mut self, config: crate::config::Config) -> Self {
-        if self.report_mode.is_none() {
-            self.report_mode = Some(config.common.report_mode);
-        }
-        self
-    }
-}
-
 /// The main function for the `validate` subcommand.
-pub async fn validate(args: Args) -> CommandResult<()> {
+pub async fn validate(args: Args, config: Config) -> CommandResult<()> {
     if let Source::Directory(_) = args.source {
         return Err(
             anyhow!("directory sources are not supported for the `validate` command").into(),
@@ -68,6 +59,7 @@ pub async fn validate(args: Args) -> CommandResult<()> {
 
     let results = Analysis::default()
         .add_source(args.source.clone())
+        .fallback_version(config.common.wdl.fallback_version)
         .run()
         .await
         .map_err(CommandError::from)?;
@@ -86,7 +78,7 @@ pub async fn validate(args: Args) -> CommandResult<()> {
         })?
         .into_engine_invocation(document)?;
 
-    let (name, inputs, _) = if let Some(inputs) = inputs {
+    let (name, inputs, origins) = if let Some(inputs) = inputs {
         inputs
     } else {
         // No inputs provided
@@ -127,15 +119,31 @@ pub async fn validate(args: Args) -> CommandResult<()> {
     };
 
     match inputs {
-        EngineInputs::Task(inputs) => {
+        EngineInputs::Task(mut inputs) => {
             // SAFETY: we wouldn't have a task inputs if a task didn't exist
             // that matched the user's criteria.
-            inputs.validate(document, document.task_by_name(&name).unwrap(), None)?
+            let task = document.task_by_name(&name).unwrap();
+            inputs
+                .join_paths(task, |key| {
+                    origins
+                        .get(key)
+                        .ok_or(anyhow!("unable to find origin path for key `{key}`"))
+                })
+                .await?;
+            inputs.validate(document, task, None)?
         }
-        EngineInputs::Workflow(inputs) => {
+        EngineInputs::Workflow(mut inputs) => {
             // SAFETY: we wouldn't have a workflow inputs if a workflow didn't
             // exist that matched the user's criteria.
-            inputs.validate(document, document.workflow().unwrap(), None)?
+            let workflow = document.workflow().unwrap();
+            inputs
+                .join_paths(workflow, |key| {
+                    origins
+                        .get(key)
+                        .ok_or(anyhow!("unable to find origin path for key `{key}`"))
+                })
+                .await?;
+            inputs.validate(document, workflow, None)?
         }
     }